[feat][yunnij] - news_analyzer.py 통합
