+++ app/main.py
@@ -3,8 +3,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from pydantic import BaseModel, Field
 from typing import List, Optional, Dict
 
-from app.free_news_similarity_analyzer import free_analyze_and_summarize
-from app.premium_news_similarity_analyzer import premium_analyze_and_summarize
+from app.news_analyzer import analyze_and_summarize
 
 app = FastAPI(title="뉴스 기사 유사도 분석 및 요약 API", version="1.0.0")
 
@@ -27,7 +26,7 @@ class News(BaseModel):
 
 class NewsDto(BaseModel):
     plan: str = Field(default="NORMAL", example = "NORMAL")
-    category: Optional[str] = Field(default=None, example="IT/과학") 
+    category: Optional[str] = Field(default=None, example="IT/과학", description="뉴스 카테고리")
     mainNewsDto: News = Field(
         example={
             "url": "https://www.e-platform.net/news/articleView.html?idxno=84363",
@@ -35,7 +34,7 @@ class NewsDto(BaseModel):
             "date": "2024-03-26",
             "content": "▲ 재생원료 인증 시범사업 추진(안) [에너지플랫폼뉴스 송승온 기자] 정부가 ‘전기차 폐배터리 재생원료 인증 시범사업’을 통해 인증제도 마련 및 관련 제도 개선을 추진한다. 환경부는 26일 엘더블유(LW)컨벤션센터에서 5개 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 체결했다고 밝혔다. 이날 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석했다. 유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다. 특히 폐배터리를 재활용해 재생원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다. 환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다. 이번에 참여한 5개 기업의 시범사업을 통해 한국환경공단에서 운영하는 폐기물관리시스템인 ‘올바로 시스템’을 활용, 재생원료 추출 등 재활용 공정 과정을 검증한다. ‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용해 실시간으로 투명하게 관리하기 위한 종합관리시스템이다. 재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다. 환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다. 아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다. 한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 밝혔다.",
             "image": "https://cdn.e-platform.net/news/photo/202403/84363_59942_577.jpg"
-        })
+        }, description="주 뉴스 데이터")
     newsDtos: List[News] = Field(
         example=[
             {
@@ -50,7 +49,7 @@ class NewsDto(BaseModel):
                 "date": "2024-03-26",
                 "content": "신문구독신청 로그인 회원가입 Update2025.07.19(토) 뉴스종합 전체 종합 경제 소재 부품 로봇/장비 산업가스 수소 탄소 배터리 3D프린팅 전자 에너지 환경안전/소방 테크노파크 농림·산림 국토교통 칼럼·기고 신소재칼럼 기고 기획·인터뷰 기획 인터뷰 독자마당 자유게시판 공지사항 포토갤러리 뉴스레터 유료기사 구독신청 기사 메일전송 홈 뉴스종합 배터리 환경부, 전기차 폐배터리 재생원료 인증 시범사업 추진 - 성일하이텍 등 5社, 니켈·코발트 등 핵심광물 공급 페이스북 공유하기 트위터 공유하기 구글플러스 공유하기 싸이월드 공유하기 라인 공유하기 네이버블로그 공유하기 네이버밴드 공유하기 네이트온쪽지 공유하기 구글북마크 공유하기 스크랩하기 프린트하기 이메일보내기 글자확대 글자축소 기사등록 2024-03-26 17:31:13 수정 2024-03-27 16:59:38 기사수정 정부가 전기차 폐배터리 재생원료 인증을 통해 국내 재활용사 및 배터리제조사의 수출입 지원 과 글로벌 무역규제 대응에 나선다. 환경부(장관 한화진)는 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 26일 체결한다고 밝혔다. 이번 시범사업에 참여하는 기업은 △성일하이텍 △에코프로씨엔지 △포스코HY클린메탈 △SK에코플랜트 △에너지머티리얼즈 등 5개 기업이다. 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석한다. 유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다. 특히 폐배터리를 재활용해 재생 원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다. 환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다. 이번에 참여한 5개 기업의 시범사업은 한국환경공단에서 운영하는 폐기물관리 시스템인 ‘올바로 시스템’을 활용해 재생원료 추출 등 재활용 공정 과정을 검증한다. ‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용하여 실시간으로 투명하게 관리하기 위한 종합관리시스템이다. 재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다. 환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다. 아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다. 한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기 전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 전했다. 0 기사수정 다른 곳에 퍼가실 때는 아래 고유 링크 주소를 출처로 사용해주세요. http://amenews.kr/news/view.php?idx=56907 기사등록 2024-03-26 17:31:13 수정 2024-03-27 16:59:38 페이스북 공유하기 트위터 공유하기 구글플러스 공유하기 싸이월드 공유하기 라인 공유하기 네이버블로그 공유하기 네이버밴드 공유하기 네이트온쪽지 공유하기 구글북마크 공유하기 스크랩하기 프린트하기 이메일보내기 글자확대 글자축소 기자프로필 유혜리 기자 유혜리 기자의 다른 기사 보기 나도 한마디 ※ 로그인 후 의견을 등록하시면, 자신의 의견을 관리하실 수 있습니다. 0/1000 　많이 본 뉴스 1 엔알비, 공모가 21000원 확정 2 원익IPS, 171억 E 수요관리 기술개발 주관기관 3 中 실리콘 음극재 특허 출원 68%, 독주 체제 4 에이엠솔루션즈, 대형 부품 아크 와이어 3DP 선 5 가전·車 업계, 신제품 개발 3DP 적용 비용·시간 ↓ 6 동서발전, 천연가스 발전 포집 탄산 출하 7 PBV 글로벌 수요 증가, 韓 車 수출 기회 8 SEMI, 4년간 첨단 반도체 생산능력 69%↑ 전망 9 서울과기대, 유연 메커니즘 최적설계기술 개발 10 中 리튬價, '22년 고점比 90%↓ 개인정보취급방침 이용약관 메일수집거부 제휴제안 광고문의 청소년보호정책 찾아오시는길 신소재경제신문 등록번호서울특별시 다09721 등록일2009-07-01 발행일자2025-07-19 발행인고봉길 편집인고봉길 청소년보호책임자고봉길 연락처02-2055-1632 이메일webmaster@amenews.kr 주소서울특별시 관악구 보라매5가길7 캐릭터그린빌 309호 신소재경제신문 © amenews.kr All rights reserved. 신소재경제신문의 모든 콘텐츠(기사 등)는 저작권법의 보호를 받은바, 무단 전재, 복사, 배포 등을 금합니다. 모바일 버전 바로가기"
             }
-        ])
+        ], description="비교 뉴스 리스트")
         
 
 @app.post("/analyze", description="뉴스 데이터를 분석하고 요약 결과 반환합니다.",
@@ -61,77 +60,125 @@ class NewsDto(BaseModel):
                       "application/json" :{
                         "examples": {
                         "NORMAL":{
-                            "value":{
-                            "category": "IT/과학",
-                            "mainNewsDto": {
-                                "url": "https://www.e-platform.net/news/articleView.html?idxno=84363",
-                                "title": "폐배터리 재생원료 인증사업 추진… 재활용 공정과정 검증 < 종합 < 기사본문",
-                                "date": "2024-03-26",
-                                "content": "▲ 재생원료 인증 시범사업 추진(안) [에너지플랫폼뉴스 송승온 기자] 정부가 ‘전기차 폐배터리 재생원료 인증 시범사업’을 통해 인증제도 마련 및 관련 제도 개선을 추진한다. 환경부는 26일 엘더블유(LW)컨벤션센터에서 5개 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 체결했다고 밝혔다. 이날 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석했다. 유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다. 특히 폐배터리를 재활용해 재생원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다. 환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다. 이번에 참여한 5개 기업의 시범사업을 통해 한국환경공단에서 운영하는 폐기물관리시스템인 ‘올바로 시스템’을 활용, 재생원료 추출 등 재활용 공정 과정을 검증한다. ‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용해 실시간으로 투명하게 관리하기 위한 종합관리시스템이다. 재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다. 환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다. 아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다. 한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 밝혔다.",
-                                "image": "https://cdn.e-platform.net/news/photo/202403/84363_59942_577.jpg"
-                            },
-                            "newsComparisonDtos": [
-                                {
-                                "newsWithSimilarityDto": {
-                                    "similarity": 1,
-                                    "newsDto": {
-                                    "url": "http://amenews.kr/news/view.php?idx=56907",
-                                    "title": "<b>환경부</b>, <b>전기차 폐배터리 재생원료 인증 시범사업</b> 추진",
-                                    "date": "2025-06-27",
-                                    "content": "신문구독신청 로그인 회원가입 Update2025.07.19(토) 뉴스종합 전체 종합 경제 소재 부품 로봇/장비 산업가스 수소 탄소 배터리 3D프린팅 전자 에너지 환경안전/소방 테크노파크 농림·산림 국토교통 칼럼·기고 신소재칼럼 기고 기획·인터뷰 기획 인터뷰 독자마당 자유게시판 공지사항 포토갤러리 뉴스레터 유료기사 구독신청 기사 메일전송 홈 뉴스종합 배터리 환경부, 전기차 폐배터리 재생원료 인증 시범사업 추진 - 성일하이텍 등 5社, 니켈·코발트 등 핵심광물 공급 페이스북 공유하기 트위터 공유하기 구글플러스 공유하기 싸이월드 공유하기 라인 공유하기 네이버블로그 공유하기 네이버밴드 공유하기 네이트온쪽지 공유하기 구글북마크 공유하기 스크랩하기 프린트하기 이메일보내기 글자확대 글자축소 기사등록 2024-03-26 17:31:13 수정 2024-03-27 16:59:38 기사수정 정부가 전기차 폐배터리 재생원료 인증을 통해 국내 재활용사 및 배터리제조사의 수출입 지원 과 글로벌 무역규제 대응에 나선다. 환경부(장관 한화진)는 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 26일 체결한다고 밝혔다. 이번 시범사업에 참여하는 기업은 △성일하이텍 △에코프로씨엔지 △포스코HY클린메탈 △SK에코플랜트 △에너지머티리얼즈 등 5개 기업이다. 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석한다. 유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다. 특히 폐배터리를 재활용해 재생 원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다. 환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다. 이번에 참여한 5개 기업의 시범사업은 한국환경공단에서 운영하는 폐기물관리 시스템인 ‘올바로 시스템’을 활용해 재생원료 추출 등 재활용 공정 과정을 검증한다. ‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용하여 실시간으로 투명하게 관리하기 위한 종합관리시스템이다. 재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다. 환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다. 아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다. 한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기 전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 전했다. 0 기사수정 다른 곳에 퍼가실 때는 아래 고유 링크 주소를 출처로 사용해주세요. http://amenews.kr/news/view.php?idx=56907 기사등록 2024-03-26 17:31:13 수정 2024-03-27 16:59:38 페이스북 공유하기 트위터 공유하기 구글플러스 공유하기 싸이월드 공유하기 라인 공유하기 네이버블로그 공유하기 네이버밴드 공유하기 네이트온쪽지 공유하기 구글북마크 공유하기 스크랩하기 프린트하기 이메일보내기 글자확대 글자축소 기자프로필 유혜리 기자 유혜리 기자의 다른 기사 보기 나도 한마디 ※ 로그인 후 의견을 등록하시면, 자신의 의견을 관리하실 수 있습니다. 0/1000 　많이 본 뉴스 1 엔알비, 공모가 21000원 확정 2 원익IPS, 171억 E 수요관리 기술개발 주관기관 3 中 실리콘 음극재 특허 출원 68%, 독주 체제 4 에이엠솔루션즈, 대형 부품 아크 와이어 3DP 선 5 가전·車 업계, 신제품 개발 3DP 적용 비용·시간 ↓ 6 동서발전, 천연가스 발전 포집 탄산 출하 7 PBV 글로벌 수요 증가, 韓 車 수출 기회 8 SEMI, 4년간 첨단 반도체 생산능력 69%↑ 전망 9 서울과기대, 유연 메커니즘 최적설계기술 개발 10 中 리튬價, '22년 고점比 90%↓ 개인정보취급방침 이용약관 메일수집거부 제휴제안 광고문의 청소년보호정책 찾아오시는길 신소재경제신문 등록번호서울특별시 다09721 등록일2009-07-01 발행일자2025-07-19 발행인고봉길 편집인고봉길 청소년보호책임자고봉길 연락처02-2055-1632 이메일webmaster@amenews.kr 주소서울특별시 관악구 보라매5가길7 캐릭터그린빌 309호 신소재경제신문 © amenews.kr All rights reserved. 신소재경제신문의 모든 콘텐츠(기사 등)는 저작권법의 보호를 받은바, 무단 전재, 복사, 배포 등을 금합니다. 모바일 버전 바로가기"
-                                    }
-                                }
-                                },
-                                {
-                                "newsWithSimilarityDto": {
-                                    "similarity": 0.853,
-                                    "newsDto": {
-                                    "url": "https://www.mediapen.com/news/view/1026344",
-                                    "title": "“<b>폐배터리</b>에 미래 캔다&quot;…EV 배터리 재활용 시장 가파른 성장세",
-                                    "date": "2024-03-26",
-                                    "content": "검색 2025-07-19 토요일 홈 경제 경제일반 금융 증권 산업 자동차/항공 조선/철강 화학/에너지 제약/바이오 중기 전자 게임 유통 부동산 칼럼 정치 사회 연예 최신 영화 연예 스포츠 스포츠종합 축구 야구 골프 여행 문화 “폐배터리에 미래 캔다\"…EV 배터리 재활용 시장 가파른 성장세 2025-06-27 14:50:58 | 박재훈 기자 | pak1005@mediapen.com 페이스북 트위터 카카오톡 프린트 가 + 가 - 국내 시장 2032년까지 연평균 6% 성장…전기차 보급 확대 주효 국내 주요 기업 기술 혁신 및 인프라 확충… 정부 지원까지 적극 확대 [미디어펜=박재훈 기자]전기차 보급이 점차 확대되면서 국내 전기차 배터리 재활용 시장도 급격한 성장세를 보인다. 사용 후 배터리의 처리와 재활용이 산업계와 정부 및 환경 분야에서 핵심 이슈로 부상하면서 순환경제 실현과 자원 확보, 글로벌 경쟁력 강화에 초점을 맞추며 관련 시장 육성에 박차를 가하고 있다. ▲ SK에코플랜트 자회사 SK테스의 네덜란드 로테르담에 폐배터리 재활용 전처리 공장 모습./사진=SK에코플랜트 27일 시장조사기관 엑스퍼트마켓리서치에 따르면 국내 시장의 배터리 재활용 시장은 향후 10년 간 약 70%의 성장세를 보일 것으로 전망된다. 글로벌 기준으로 전기차 배터리 재활용 시장은 올해 기준 약 49억 달러 규모로 평가되며 오는 2034년에는 428억 달러까지 성장세를 이어갈 것으로 예상된다. 전문가들은 국내 사용후 배터리 재활용 시장이 2032년까지 3억4400만 달러(약 4655억 원) 규모로 성장할 것으로 보고 있다. 연평균 성장률은 약 6%로 글로벌 시장의 고성장세(연평균 27.3%)에 비하면 다소 낮은 수치지만 국내 산업 구조와 정책 환경을 감안하면 의미 있는 성장세라는 평가다. 전기차와 에너지저장장치(ESS) 등 친환경 모빌리티와 신재생에너지 산업이 확산되면서 수명이 다한 배터리의 재활용 수요도 함께 급증하고 있다. 국내 배터리 재활용 시장의 성장 동력은 크게 △전기차 보급 확대 △환경 규제와 정책 강화 △자원 확보와 공급망 안정성 △기술 혁신과 인프라 확충 네 가지로 요약된다. 전기차 보급 확대는 가장 큰 요인으로 평가된다. 2024년 5월 기준 국내 전기차 누적 등록대수는 59만 대를 넘어섰다. 앞서 환경부는 2030년 이후 연간 10만 개 이상의 사용후배터리가 배출될 것으로 내다봤다. 이를 위해 환경부는 판매되는 전기차 외에도 이후 폐배터리 회수를 위해 배터리 환경성 계수를 적용하기도 했다. 보조금에 차등 적용을 유인했으며 전기차 메이커들에게 핵심 광물 회수율이 높은 NCM(니켈, 코발트, 망간) 배터리 사용을 적극 권장했다. 둘째는 환경 규제와 정책 강화다. 정부는 2025년까지 재생원료 인증제를 도입하고 2027년부터 배터리 등급 분류를 위한 성능평가를 시행할 계획이다. 해당 제도는 사용후배터리에서 추출한 리튬, 니켈 등 유가금속이 신품 배터리 제조에 얼마나 투입됐는지 확인하는 체계로 EU 등 주요국의 통상규제에 대응하고 국내 기업의 수출 경쟁력도 높일 것으로 기대된다. 자원 확보와 공급망 안정성도 주효한 이유다. 국내 배터리 산업은 리튬, 코발트, 니켈 등 핵심 광물의 대부분을 수입에 의존하고 있는 상황이다. 배터리 재활용을 통해 금속들을 회수할 경우 원가 경쟁력과 공급망 안정성을 동시에 확보할 수 있다는 장점이 있다. 환경부는 지난해 3월 국내 재활용 기업들과 ‘전기차 폐배터리 재생원료 인증 시범사업’ 협약을 체결했다. 해당 사업은 폐배터리에서 추출한 재생원료의 품질과 이력을 인증하는 제도로 순환경제 실현과 희소 금속 공급망 강화에 기여하기 위한 취지다. ▲ 성일하이텍 직원들이 폐 배터리 해체 작업을 하고 있다./사진=KBS1 시사멘터리 추적 '폐배터리가 몰려온다' 캡처 기술 혁신과 인프라 확충도 폐배터리 시장의 규모를 키우는 요인으로 꼽힌다. 국내 주요 기업들은 습식제련, 건식제련, 직접 재활용 등 다양한 기술을 개발·적용하고 있다. 성일하이텍, 에코프로, 포스코HY클린메탈, SK에코플랜트, 영풍, 고려아연 등은 배터리 전처리(분해·파쇄)부터 후처리(금속 추출·정제)까지 밸류체인을 내재화하며 경쟁력을 높이고 있다. 특히 에코프로는 포항에 2조 원을 투입해 폐배터리 재활용 시스템을 구축했다. 기업들의 노력을 장려하기 위해 정부의 지원도 확대되고 있다. 환경부와 한국환경공단은 민간의 전기차 폐배터리 회수·관리체계 구축을 위해 분리, 보관, 화재방지, 성능평가, 방전 설비 구매비용의 50%를 지원하는 사업을 시행 중이다. 사업자당 최대 1억 원의 지원금이 지급되며 이를 통해 민간 회수체계와 안전한 산업 환경 조성, 재사용·재활용 활성화가 기대된다. 국내 배터리 재활용 산업은 글로벌 경쟁 심화, 규제 변화, 공급망 차질 등 과제도 안고 있다. 그러나 정부와 업계의 협력, 기술 개발, 인프라 확충, 정책 지원이 맞물리며 시장은 지속적으로 성장할 것으로 전망된다. 이를 위해 환경부는 생산자책임재활용제도(EPR) 대상을 기존 50종에서 전 품목으로 확대했다. 전기·전자제품 내 배터리 회수율을 극대화하고 재활용이 까다로운 LFP(리튬, 인산, 철) 배터리 등도 EPR 대상에 포함해 공급망 차질에 적극 대응하고 있다. 또한 전국 6개소에 있는 재활용 가능 자원 비축시설(약 1만5000톤 규모)을 원료 보관 장소로 기업에 임대해 원료 수급의 안정성을 높이고 폐기물 규제 완화로 순환자원 인정 범위도 확대하고 있다. 환경부는 \"이 같은 대책을 계기로 국내 배터리 순환이용 업계가 향후 2~3년 내에 다가올 산업 성장기에도 시장 경제를 지속 유지할 수 있도록 적극 지원할 것\"이라고 말했다. [미디어펜=박재훈 기자] ▶다른기사보기 <저작권자 © 미디어펜 무단전재 및 재배포금지> 라이브리 댓글 작성을 위해 JavaScript를 활성화 해주세요 회사소개 연혁 광고·제휴문의 청소년보호정책 이메일무단수집거부 디지털 뉴스콘텐츠 이용규칙 서울특별시 종로구 새문안로3길 30 세종로대우빌딩 복합동 508호 전화번호: 02)6241-7700 팩스: 02)6241-7708 정기간행물ㆍ등록번호: 서울 아 00574 등록일.발행일 2008.5.8 발행인: 이의춘 편집인·편집국장: 김태균 청소년보호책임자: 김사성 Copyright © 2013 미디어펜. All rights reserved."
-                                    }
-                                }
-                                }
-                            ]
-                        }},
+                            "value":
+{
+  "category": "IT/과학",
+  "plan": "NORMAL",
+  "mainNewsDto": {
+    "url": "https://www.e-platform.net/news/articleView.html?idxno=84363",
+    "title": "폐배터리 재생원료 인증사업 추진… 재활용 공정과정 검증 < 종합 < 기사본문",
+    "date": "2024-03-26",
+    "content": "환경부는 26일 엘더블유(LW)컨벤션센터에서 5개 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 체결했다고 밝혔다.\n이날 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석했다.\n유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다.\n특히 폐배터리를 재활용해 재생원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다.\n환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다.\n이번에 참여한 5개 기업의 시범사업을 통해 한국환경공단에서 운영하는 폐기물관리시스템인 ‘올바로 시스템’을 활용, 재생원료 추출 등 재활용 공정 과정을 검증한다.\n‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용해 실시간으로 투명하게 관리하기 위한 종합관리시스템이다.\n재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다.\n환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다.\n아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다.\n한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 밝혔다.",
+    "image": "https://cdn.e-platform.net/news/photo/202403/84363_59942_577.jpg",
+    "phrases": [
+      "이날 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석했다.",
+      "환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다.",
+      "아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다.",
+      "한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 밝혔다."
+    ]
+  },
+  "newsComparisonDtos": [
+    {
+      "newsWithSimilarityDto": {
+        "similarity": 1,
+        "newsDto": {
+          "url": "http://amenews.kr/news/view.php?idx=56907",
+          "title": "<b>환경부</b>, <b>전기차 폐배터리 재생원료 인증 시범사업</b> 추진",
+          "date": "2024-03-26",
+          "content": "환경부(장관 한화진)는 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 26일 체결한다고 밝혔다.\n이번 시범사업에 참여하는 기업은 △성일하이텍 △에코프로씨엔지 △포스코HY클린메탈 △SK에코플랜트 △에너지머티리얼즈 등 5개 기업이다.\n협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석한다.\n유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다.\n특히 폐배터리를 재활용해 재생 원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다.\n환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다.\n이번에 참여한 5개 기업의 시범사업은 한국환경공단에서 운영하는 폐기물관리 시스템인 ‘올바로 시스템’을 활용해 재생원료 추출 등 재활용 공정 과정을 검증한다.\n‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용하여 실시간으로 투명하게 관리하기 위한 종합관리시스템이다.\n재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다.\n환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다.\n아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다.\n한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기 전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 전했다.\n신소재경제신문의 모든 콘텐츠(기사 등)는 저작권법의 보호를 받은바, 무단 전재, 복사, 배포 등을 금합니다.\n모바일 버전 바로가기",
+          "phrases": [
+            "환경부(장관 한화진)는 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 26일 체결한다고 밝혔다.",
+            "협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석한다.",
+            "환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다.",
+            "아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다.",
+            "한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기 전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 전했다."
+          ]
+        }
+      },
+      "comparison": "메인 기사와 비교 기사 모두 환경부가 전기차 폐배터리 재생원료 인증 시범사업을 추진하고 있다는 점과 5개 기업의 참여를 강조하고 있습니다. 그러나 메인 기사에서는 '올바로 시스템'의 활용과 재활용 공정 검증에 초점을 맞추고 있는 반면, 비교 기사에서는 인증이 국내 재활용사 및 배터리 제조사의 수출입 지원과 글로벌 무역규제 대응에 기여할 것이라고 언급하고 있습니다."
+    },
+    {
+      "newsWithSimilarityDto": {
+        "similarity": 0.853,
+        "newsDto": {
+          "url": "https://www.mediapen.com/news/view/1026344",
+          "title": "“<b>폐배터리</b>에 미래 캔다&quot;…EV 배터리 재활용 시장 가파른 성장세",
+          "date": "2025-06-27",
+          "content": "사용 후 배터리의 처리와 재활용이 산업계와 정부 및 환경 분야에서 핵심 이슈로 부상하면서 순환경제 실현과 자원 확보, 글로벌 경쟁력 강화에 초점을 맞추며 관련 시장 육성에 박차를 가하고 있다.\n▲ SK에코플랜트 자회사 SK테스의 네덜란드 로테르담에 폐배터리 재활용 전처리 공장 모습./사진=SK에코플랜트 27일 시장조사기관 엑스퍼트마켓리서치에 따르면 국내 시장의 배터리 재활용 시장은 향후 10년 간 약 70%의 성장세를 보일 것으로 전망된다.\n글로벌 기준으로 전기차 배터리 재활용 시장은 올해 기준 약 49억 달러 규모로 평가되며 오는 2034년에는 428억 달러까지 성장세를 이어갈 것으로 예상된다.\n전문가들은 국내 사용후 배터리 재활용 시장이 2032년까지 3억4400만 달러(약 4655억 원) 규모로 성장할 것으로 보고 있다.\n연평균 성장률은 약 6%로 글로벌 시장의 고성장세(연평균 27.3%)에 비하면 다소 낮은 수치지만 국내 산업 구조와 정책 환경을 감안하면 의미 있는 성장세라는 평가다.\n전기차와 에너지저장장치(ESS) 등 친환경 모빌리티와 신재생에너지 산업이 확산되면서 수명이 다한 배터리의 재활용 수요도 함께 급증하고 있다.\n국내 배터리 재활용 시장의 성장 동력은 크게 △전기차 보급 확대 △환경 규제와 정책 강화 △자원 확보와 공급망 안정성 △기술 혁신과 인프라 확충 네 가지로 요약된다.\n전기차 보급 확대는 가장 큰 요인으로 평가된다.\n2024년 5월 기준 국내 전기차 누적 등록대수는 59만 대를 넘어섰다.\n앞서 환경부는 2030년 이후 연간 10만 개 이상의 사용후배터리가 배출될 것으로 내다봤다.\n이를 위해 환경부는 판매되는 전기차 외에도 이후 폐배터리 회수를 위해 배터리 환경성 계수를 적용하기도 했다.\n보조금에 차등 적용을 유인했으며 전기차 메이커들에게 핵심 광물 회수율이 높은 NCM(니켈, 코발트, 망간) 배터리 사용을 적극 권장했다.\n둘째는 환경 규제와 정책 강화다.\n정부는 2025년까지 재생원료 인증제를 도입하고 2027년부터 배터리 등급 분류를 위한 성능평가를 시행할 계획이다.\n해당 제도는 사용후배터리에서 추출한 리튬, 니켈 등 유가금속이 신품 배터리 제조에 얼마나 투입됐는지 확인하는 체계로 EU 등 주요국의 통상규제에 대응하고 국내 기업의 수출 경쟁력도 높일 것으로 기대된다.\n자원 확보와 공급망 안정성도 주효한 이유다.\n국내 배터리 산업은 리튬, 코발트, 니켈 등 핵심 광물의 대부분을 수입에 의존하고 있는 상황이다.\n배터리 재활용을 통해 금속들을 회수할 경우 원가 경쟁력과 공급망 안정성을 동시에 확보할 수 있다는 장점이 있다.\n환경부는 지난해 3월 국내 재활용 기업들과 ‘전기차 폐배터리 재생원료 인증 시범사업’ 협약을 체결했다.\n해당 사업은 폐배터리에서 추출한 재생원료의 품질과 이력을 인증하는 제도로 순환경제 실현과 희소 금속 공급망 강화에 기여하기 위한 취지다.\n▲ 성일하이텍 직원들이 폐 배터리 해체 작업을 하고 있다./사진=KBS1 시사멘터리 추적 '폐배터리가 몰려온다' 캡처 기술 혁신과 인프라 확충도 폐배터리 시장의 규모를 키우는 요인으로 꼽힌다.\n국내 주요 기업들은 습식제련, 건식제련, 직접 재활용 등 다양한 기술을 개발·적용하고 있다.\n성일하이텍, 에코프로, 포스코HY클린메탈, SK에코플랜트, 영풍, 고려아연 등은 배터리 전처리(분해·파쇄)부터 후처리(금속 추출·정제)까지 밸류체인을 내재화하며 경쟁력을 높이고 있다.\n특히 에코프로는 포항에 2조 원을 투입해 폐배터리 재활용 시스템을 구축했다.\n기업들의 노력을 장려하기 위해 정부의 지원도 확대되고 있다.\n환경부와 한국환경공단은 민간의 전기차 폐배터리 회수·관리체계 구축을 위해 분리, 보관, 화재방지, 성능평가, 방전 설비 구매비용의 50%를 지원하는 사업을 시행 중이다.\n사업자당 최대 1억 원의 지원금이 지급되며 이를 통해 민간 회수체계와 안전한 산업 환경 조성, 재사용·재활용 활성화가 기대된다.\n국내 배터리 재활용 산업은 글로벌 경쟁 심화, 규제 변화, 공급망 차질 등 과제도 안고 있다.\n그러나 정부와 업계의 협력, 기술 개발, 인프라 확충, 정책 지원이 맞물리며 시장은 지속적으로 성장할 것으로 전망된다.\n이를 위해 환경부는 생산자책임재활용제도(EPR) 대상을 기존 50종에서 전 품목으로 확대했다.\n전기·전자제품 내 배터리 회수율을 극대화하고 재활용이 까다로운 LFP(리튬, 인산, 철) 배터리 등도 EPR 대상에 포함해 공급망 차질에 적극 대응하고 있다.\n또한 전국 6개소에 있는 재활용 가능 자원 비축시설(약 1만5000톤 규모)을 원료 보관 장소로 기업에 임대해 원료 수급의 안정성을 높이고 폐기물 규제 완화로 순환자원 인정 범위도 확대하고 있다.\n환경부는 \"이 같은 대책을 계기로 국내 배터리 순환이용 업계가 향후 2~3년 내에 다가올 산업 성장기에도 시장 경제를 지속 유지할 수 있도록 적극 지원할 것\"이라고 말했다.",
+          "phrases": [
+            "전문가들은 국내 사용후 배터리 재활용 시장이 2032년까지 3억4400만 달러(약 4655억 원) 규모로 성장할 것으로 보고 있다.",
+            "정부는 2025년까지 재생원료 인증제를 도입하고 2027년부터 배터리 등급 분류를 위한 성능평가를 시행할 계획이다.",
+            "사용 후 배터리의 처리와 재활용이 산업계와 정부 및 환경 분야에서 핵심 이슈로 부상하면서 순환경제 실현과 자원 확보, 글로벌 경쟁력 강화에 초점을 맞추며 관련 시장 육성에 박차를 가하고 있다.",
+            "▲ SK에코플랜트 자회사 SK테스의 네덜란드 로테르담에 폐배터리 재활용 전처리 공장 모습./사진=SK에코플랜트 27일 시장조사기관 엑스퍼트마켓리서치에 따르면 국내 시장의 배터리 재활용 시장은 향후 10년 간 약 70%의 성장세를 보일 것으로 전망된다.",
+            "기업들의 노력을 장려하기 위해 정부의 지원도 확대되고 있다."
+          ]
+        }
+      },
+      "comparison": "메인 기사에서는 전기차 폐배터리의 재생원료 인증 시범사업을 통해 인증제도 마련과 제도 개선을 추진하는 정부의 노력을 강조하고 있는 반면, 비교 기사에서는 전기차 배터리 재활용 시장의 급격한 성장세와 그에 따른 국내 시장의 전망을 중심으로 다루고 있습니다."
+    }
+  ]
+}
+                        },
                         "PREMIUM":{
-                            "value":{
-                            "category": "IT/과학",
-                            "mainNewsDto": {
-                                "url": "https://www.e-platform.net/news/articleView.html?idxno=84363",
-                                "title": "폐배터리 재생원료 인증사업 추진… 재활용 공정과정 검증 < 종합 < 기사본문",
-                                "date": "2024-03-26",
-                                "content": "▲ 재생원료 인증 시범사업 추진(안) [에너지플랫폼뉴스 송승온 기자] 정부가 ‘전기차 폐배터리 재생원료 인증 시범사업’을 통해 인증제도 마련 및 관련 제도 개선을 추진한다. 환경부는 26일 엘더블유(LW)컨벤션센터에서 5개 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 체결했다고 밝혔다. 이날 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석했다. 유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다. 특히 폐배터리를 재활용해 재생원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다. 환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다. 이번에 참여한 5개 기업의 시범사업을 통해 한국환경공단에서 운영하는 폐기물관리시스템인 ‘올바로 시스템’을 활용, 재생원료 추출 등 재활용 공정 과정을 검증한다. ‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용해 실시간으로 투명하게 관리하기 위한 종합관리시스템이다. 재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다. 환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다. 아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다. 한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 밝혔다.",
-                                "image": "https://cdn.e-platform.net/news/photo/202403/84363_59942_577.jpg"
-                            },
-                            "newsComparisonDtos": [
-                                {
-                                "newsWithSimilarityDto": {
-                                    "similarity": 1,
-                                    "newsDto": {
-                                    "url": "http://amenews.kr/news/view.php?idx=56907",
-                                    "title": "<b>환경부</b>, <b>전기차 폐배터리 재생원료 인증 시범사업</b> 추진",
-                                    "date": "2025-06-27",
-                                    "content": "신문구독신청 로그인 회원가입 Update2025.07.19(토) 뉴스종합 전체 종합 경제 소재 부품 로봇/장비 산업가스 수소 탄소 배터리 3D프린팅 전자 에너지 환경안전/소방 테크노파크 농림·산림 국토교통 칼럼·기고 신소재칼럼 기고 기획·인터뷰 기획 인터뷰 독자마당 자유게시판 공지사항 포토갤러리 뉴스레터 유료기사 구독신청 기사 메일전송 홈 뉴스종합 배터리 환경부, 전기차 폐배터리 재생원료 인증 시범사업 추진 - 성일하이텍 등 5社, 니켈·코발트 등 핵심광물 공급 페이스북 공유하기 트위터 공유하기 구글플러스 공유하기 싸이월드 공유하기 라인 공유하기 네이버블로그 공유하기 네이버밴드 공유하기 네이트온쪽지 공유하기 구글북마크 공유하기 스크랩하기 프린트하기 이메일보내기 글자확대 글자축소 기사등록 2024-03-26 17:31:13 수정 2024-03-27 16:59:38 기사수정 정부가 전기차 폐배터리 재생원료 인증을 통해 국내 재활용사 및 배터리제조사의 수출입 지원 과 글로벌 무역규제 대응에 나선다. 환경부(장관 한화진)는 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 26일 체결한다고 밝혔다. 이번 시범사업에 참여하는 기업은 △성일하이텍 △에코프로씨엔지 △포스코HY클린메탈 △SK에코플랜트 △에너지머티리얼즈 등 5개 기업이다. 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석한다. 유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다. 특히 폐배터리를 재활용해 재생 원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다. 환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다. 이번에 참여한 5개 기업의 시범사업은 한국환경공단에서 운영하는 폐기물관리 시스템인 ‘올바로 시스템’을 활용해 재생원료 추출 등 재활용 공정 과정을 검증한다. ‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용하여 실시간으로 투명하게 관리하기 위한 종합관리시스템이다. 재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다. 환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다. 아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다. 한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기 전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 전했다. 0 기사수정 다른 곳에 퍼가실 때는 아래 고유 링크 주소를 출처로 사용해주세요. http://amenews.kr/news/view.php?idx=56907 기사등록 2024-03-26 17:31:13 수정 2024-03-27 16:59:38 페이스북 공유하기 트위터 공유하기 구글플러스 공유하기 싸이월드 공유하기 라인 공유하기 네이버블로그 공유하기 네이버밴드 공유하기 네이트온쪽지 공유하기 구글북마크 공유하기 스크랩하기 프린트하기 이메일보내기 글자확대 글자축소 기자프로필 유혜리 기자 유혜리 기자의 다른 기사 보기 나도 한마디 ※ 로그인 후 의견을 등록하시면, 자신의 의견을 관리하실 수 있습니다. 0/1000 　많이 본 뉴스 1 엔알비, 공모가 21000원 확정 2 원익IPS, 171억 E 수요관리 기술개발 주관기관 3 中 실리콘 음극재 특허 출원 68%, 독주 체제 4 에이엠솔루션즈, 대형 부품 아크 와이어 3DP 선 5 가전·車 업계, 신제품 개발 3DP 적용 비용·시간 ↓ 6 동서발전, 천연가스 발전 포집 탄산 출하 7 PBV 글로벌 수요 증가, 韓 車 수출 기회 8 SEMI, 4년간 첨단 반도체 생산능력 69%↑ 전망 9 서울과기대, 유연 메커니즘 최적설계기술 개발 10 中 리튬價, '22년 고점比 90%↓ 개인정보취급방침 이용약관 메일수집거부 제휴제안 광고문의 청소년보호정책 찾아오시는길 신소재경제신문 등록번호서울특별시 다09721 등록일2009-07-01 발행일자2025-07-19 발행인고봉길 편집인고봉길 청소년보호책임자고봉길 연락처02-2055-1632 이메일webmaster@amenews.kr 주소서울특별시 관악구 보라매5가길7 캐릭터그린빌 309호 신소재경제신문 © amenews.kr All rights reserved. 신소재경제신문의 모든 콘텐츠(기사 등)는 저작권법의 보호를 받은바, 무단 전재, 복사, 배포 등을 금합니다. 모바일 버전 바로가기",
-                                    "comparison": "두 기사 모두 환경부가 전기차 폐배터리 재생원료 인증 시범사업을 추진하고, 이를 통해 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제를 실현하려는 계획을 공유하고 있습니다. 하지만 메인 기사에서는 이를 통해 재생원료 인증제도를 마련하고, 기업의 부담을 최소화하며, 희소금속 공급망 안전성 강화에 기여할 계획이라고 언급하는 반면, 비교 기사에서는 국내 재활용사 및 배터리제조사의 수출입 지원과 글로벌 무역규제 대응을 강조하고 있습니다."
-                                    }
-                                }
-                                },
-                                {
-                                "newsWithSimilarityDto": {
-                                    "similarity": 0.853,
-                                    "newsDto": {
-                                    "url": "https://www.mediapen.com/news/view/1026344",
-                                    "title": "“<b>폐배터리</b>에 미래 캔다&quot;…EV 배터리 재활용 시장 가파른 성장세",
-                                    "date": "2024-03-26",
-                                    "content": "검색 2025-07-19 토요일 홈 경제 경제일반 금융 증권 산업 자동차/항공 조선/철강 화학/에너지 제약/바이오 중기 전자 게임 유통 부동산 칼럼 정치 사회 연예 최신 영화 연예 스포츠 스포츠종합 축구 야구 골프 여행 문화 “폐배터리에 미래 캔다\"…EV 배터리 재활용 시장 가파른 성장세 2025-06-27 14:50:58 | 박재훈 기자 | pak1005@mediapen.com 페이스북 트위터 카카오톡 프린트 가 + 가 - 국내 시장 2032년까지 연평균 6% 성장…전기차 보급 확대 주효 국내 주요 기업 기술 혁신 및 인프라 확충… 정부 지원까지 적극 확대 [미디어펜=박재훈 기자]전기차 보급이 점차 확대되면서 국내 전기차 배터리 재활용 시장도 급격한 성장세를 보인다. 사용 후 배터리의 처리와 재활용이 산업계와 정부 및 환경 분야에서 핵심 이슈로 부상하면서 순환경제 실현과 자원 확보, 글로벌 경쟁력 강화에 초점을 맞추며 관련 시장 육성에 박차를 가하고 있다. ▲ SK에코플랜트 자회사 SK테스의 네덜란드 로테르담에 폐배터리 재활용 전처리 공장 모습./사진=SK에코플랜트 27일 시장조사기관 엑스퍼트마켓리서치에 따르면 국내 시장의 배터리 재활용 시장은 향후 10년 간 약 70%의 성장세를 보일 것으로 전망된다. 글로벌 기준으로 전기차 배터리 재활용 시장은 올해 기준 약 49억 달러 규모로 평가되며 오는 2034년에는 428억 달러까지 성장세를 이어갈 것으로 예상된다. 전문가들은 국내 사용후 배터리 재활용 시장이 2032년까지 3억4400만 달러(약 4655억 원) 규모로 성장할 것으로 보고 있다. 연평균 성장률은 약 6%로 글로벌 시장의 고성장세(연평균 27.3%)에 비하면 다소 낮은 수치지만 국내 산업 구조와 정책 환경을 감안하면 의미 있는 성장세라는 평가다. 전기차와 에너지저장장치(ESS) 등 친환경 모빌리티와 신재생에너지 산업이 확산되면서 수명이 다한 배터리의 재활용 수요도 함께 급증하고 있다. 국내 배터리 재활용 시장의 성장 동력은 크게 △전기차 보급 확대 △환경 규제와 정책 강화 △자원 확보와 공급망 안정성 △기술 혁신과 인프라 확충 네 가지로 요약된다. 전기차 보급 확대는 가장 큰 요인으로 평가된다. 2024년 5월 기준 국내 전기차 누적 등록대수는 59만 대를 넘어섰다. 앞서 환경부는 2030년 이후 연간 10만 개 이상의 사용후배터리가 배출될 것으로 내다봤다. 이를 위해 환경부는 판매되는 전기차 외에도 이후 폐배터리 회수를 위해 배터리 환경성 계수를 적용하기도 했다. 보조금에 차등 적용을 유인했으며 전기차 메이커들에게 핵심 광물 회수율이 높은 NCM(니켈, 코발트, 망간) 배터리 사용을 적극 권장했다. 둘째는 환경 규제와 정책 강화다. 정부는 2025년까지 재생원료 인증제를 도입하고 2027년부터 배터리 등급 분류를 위한 성능평가를 시행할 계획이다. 해당 제도는 사용후배터리에서 추출한 리튬, 니켈 등 유가금속이 신품 배터리 제조에 얼마나 투입됐는지 확인하는 체계로 EU 등 주요국의 통상규제에 대응하고 국내 기업의 수출 경쟁력도 높일 것으로 기대된다. 자원 확보와 공급망 안정성도 주효한 이유다. 국내 배터리 산업은 리튬, 코발트, 니켈 등 핵심 광물의 대부분을 수입에 의존하고 있는 상황이다. 배터리 재활용을 통해 금속들을 회수할 경우 원가 경쟁력과 공급망 안정성을 동시에 확보할 수 있다는 장점이 있다. 환경부는 지난해 3월 국내 재활용 기업들과 ‘전기차 폐배터리 재생원료 인증 시범사업’ 협약을 체결했다. 해당 사업은 폐배터리에서 추출한 재생원료의 품질과 이력을 인증하는 제도로 순환경제 실현과 희소 금속 공급망 강화에 기여하기 위한 취지다. ▲ 성일하이텍 직원들이 폐 배터리 해체 작업을 하고 있다./사진=KBS1 시사멘터리 추적 '폐배터리가 몰려온다' 캡처 기술 혁신과 인프라 확충도 폐배터리 시장의 규모를 키우는 요인으로 꼽힌다. 국내 주요 기업들은 습식제련, 건식제련, 직접 재활용 등 다양한 기술을 개발·적용하고 있다. 성일하이텍, 에코프로, 포스코HY클린메탈, SK에코플랜트, 영풍, 고려아연 등은 배터리 전처리(분해·파쇄)부터 후처리(금속 추출·정제)까지 밸류체인을 내재화하며 경쟁력을 높이고 있다. 특히 에코프로는 포항에 2조 원을 투입해 폐배터리 재활용 시스템을 구축했다. 기업들의 노력을 장려하기 위해 정부의 지원도 확대되고 있다. 환경부와 한국환경공단은 민간의 전기차 폐배터리 회수·관리체계 구축을 위해 분리, 보관, 화재방지, 성능평가, 방전 설비 구매비용의 50%를 지원하는 사업을 시행 중이다. 사업자당 최대 1억 원의 지원금이 지급되며 이를 통해 민간 회수체계와 안전한 산업 환경 조성, 재사용·재활용 활성화가 기대된다. 국내 배터리 재활용 산업은 글로벌 경쟁 심화, 규제 변화, 공급망 차질 등 과제도 안고 있다. 그러나 정부와 업계의 협력, 기술 개발, 인프라 확충, 정책 지원이 맞물리며 시장은 지속적으로 성장할 것으로 전망된다. 이를 위해 환경부는 생산자책임재활용제도(EPR) 대상을 기존 50종에서 전 품목으로 확대했다. 전기·전자제품 내 배터리 회수율을 극대화하고 재활용이 까다로운 LFP(리튬, 인산, 철) 배터리 등도 EPR 대상에 포함해 공급망 차질에 적극 대응하고 있다. 또한 전국 6개소에 있는 재활용 가능 자원 비축시설(약 1만5000톤 규모)을 원료 보관 장소로 기업에 임대해 원료 수급의 안정성을 높이고 폐기물 규제 완화로 순환자원 인정 범위도 확대하고 있다. 환경부는 \"이 같은 대책을 계기로 국내 배터리 순환이용 업계가 향후 2~3년 내에 다가올 산업 성장기에도 시장 경제를 지속 유지할 수 있도록 적극 지원할 것\"이라고 말했다. [미디어펜=박재훈 기자] ▶다른기사보기 <저작권자 © 미디어펜 무단전재 및 재배포금지> 라이브리 댓글 작성을 위해 JavaScript를 활성화 해주세요 회사소개 연혁 광고·제휴문의 청소년보호정책 이메일무단수집거부 디지털 뉴스콘텐츠 이용규칙 서울특별시 종로구 새문안로3길 30 세종로대우빌딩 복합동 508호 전화번호: 02)6241-7700 팩스: 02)6241-7708 정기간행물ㆍ등록번호: 서울 아 00574 등록일.발행일 2008.5.8 발행인: 이의춘 편집인·편집국장: 김태균 청소년보호책임자: 김사성 Copyright © 2013 미디어펜. All rights reserved.",
-                                    "comparison": "메인 기사는 정부가 전기차 폐배터리 재생원료 인증 시범사업을 추진하고, 이를 통해 재활용업계와 전문가 등으로 구성된 협의체를 운영하며, 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획임을 주로 다루고 있습니다. 반면, 비교 기사는 전기차 보급이 확대되면서 전기차 배터리 재활용 시장의 성장세를 중심으로 보고하며, 이에 따른 순환경제 실현과 자원 확보, 글로벌 경쟁력 강화에 초점을 맞추고 있습니다. 두 기사 모두 전기차 폐배터리의 재활용과 관련된 주제를 다루고 있지만, 메인 기사는 정부의 정책과 제도적 측면을, 비교 기사는 시장 동향과 성장세를 강조하고 있습니다."
-                                    }
-                                }
-                                }
-                            ]
-                        }}                                
+                            "value":
+{
+  "category": "IT/과학",
+  "plan": "PREMIUM",
+  "mainNewsDto": {
+    "url": "https://www.e-platform.net/news/articleView.html?idxno=84363",
+    "title": "폐배터리 재생원료 인증사업 추진… 재활용 공정과정 검증 < 종합 < 기사본문",
+    "date": "2024-03-26",
+    "content": "환경부는 26일 엘더블유(LW)컨벤션센터에서 5개 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 체결했다고 밝혔다.\n이날 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석했다.\n유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다.\n특히 폐배터리를 재활용해 재생원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다.\n환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다.\n이번에 참여한 5개 기업의 시범사업을 통해 한국환경공단에서 운영하는 폐기물관리시스템인 ‘올바로 시스템’을 활용, 재생원료 추출 등 재활용 공정 과정을 검증한다.\n‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용해 실시간으로 투명하게 관리하기 위한 종합관리시스템이다.\n재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다.\n환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다.\n아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다.\n한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 밝혔다.",
+    "image": "https://cdn.e-platform.net/news/photo/202403/84363_59942_577.jpg",
+    "phrases": [
+      "이날 협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석했다.",
+      "환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다.",
+      "아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다.",
+      "한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 밝혔다."
+    ]
+  },
+  "newsComparisonDtos": [
+    {
+      "newsWithSimilarityDto": {
+        "similarity": 1,
+        "newsDto": {
+          "url": "http://amenews.kr/news/view.php?idx=56907",
+          "title": "<b>환경부</b>, <b>전기차 폐배터리 재생원료 인증 시범사업</b> 추진",
+          "date": "2024-03-26",
+          "content": "환경부(장관 한화진)는 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 26일 체결한다고 밝혔다.\n이번 시범사업에 참여하는 기업은 △성일하이텍 △에코프로씨엔지 △포스코HY클린메탈 △SK에코플랜트 △에너지머티리얼즈 등 5개 기업이다.\n협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석한다.\n유럽연합(EU) 등 주요 선진국은 탄소중립 실현을 위해 제품을 생산할 때 재생원료 사용을 의무화하고 있다.\n특히 폐배터리를 재활용해 재생 원료를 생산하면 대부분 수입에 의존하는 니켈, 코발트 등의 핵심광물을 안정적으로 공급하고 순환경제 실현에도 기여할 수 있다.\n환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다.\n이번에 참여한 5개 기업의 시범사업은 한국환경공단에서 운영하는 폐기물관리 시스템인 ‘올바로 시스템’을 활용해 재생원료 추출 등 재활용 공정 과정을 검증한다.\n‘올바로 시스템’은 폐기물관리법에 따라 폐기물의 배출부터 운반 및 최종처리까지 전 과정을 인터넷, 무선인식(RFID)을 활용하여 실시간으로 투명하게 관리하기 위한 종합관리시스템이다.\n재활용업체들은 폐배터리 재생원료 정보를 이미 ‘올바로시스템’에 입력하고 있기 때문에 이 시스템을 활용하면 새로운 시스템을 구축할 필요없이 기업의 부담을 최소화하면서 재생원료 인증이 가능하며, 객관적인 신뢰성을 확보할 수 있다.\n환경부는 이번 시범사업 결과를 바탕으로 재생원료 인증제도를 마련해 순환경제 실현뿐만 아니라 희소금속 공급망 안전성 강화에도 기여할 계획이다.\n아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다.\n한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기 전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 전했다.\n신소재경제신문의 모든 콘텐츠(기사 등)는 저작권법의 보호를 받은바, 무단 전재, 복사, 배포 등을 금합니다.\n모바일 버전 바로가기",
+          "phrases": [
+            "환경부(장관 한화진)는 전기차 폐배터리 재활용 기업과 함께 전기차 폐배터리 재생원료 인증 시범사업 추진 협약을 26일 체결한다고 밝혔다.",
+            "협약식에는 한화진 환경부 장관, 이강명 성일하이텍 대표, 박석회 에코프로씨엔지 대표, 임지우 포스코에이치와이(HY)클린메탈 대표, 박경일 에스케이(SK)에코플랜트 대표, 박용한 에너지머티리얼즈 상무, 정재웅 한국환경공단 이사가 참석한다.",
+            "환경부는 이번 폐배터리 재생원료 인증 시범사업을 통해 재활용업계, 전문가 등으로 구성된 협의체를 운영하고 재생원료 인증방법 등을 구체화해 인증제도 마련 및 관련 제도를 개선할 계획이다.",
+            "아울러 국내 재활용기업 및 배터리 제조사의 부담을 완화할 수 있을 것으로 기대된다.",
+            "한화진 환경부 장관은 “이번 협약을 통해 전기차 폐배터리 재활용업계와 협력하여 폐배터리 순환이용의 모범사례를 만들고, 배터리뿐만 아니라 전기 전자제품 등 여러 분야에도 재생원료 사용을 확대하여 순환경제가 사회 전반에 뿌리내릴 수 있도록 하겠다”라고 전했다."
+          ]
+        }
+      },
+      "comparison": "메인 기사와 비교 기사 모두 전기차 폐배터리 재생원료 인증 시범사업 추진과 관련된 내용을 다루고 있으며, 참여 기업과 주요 인물들이 동일하게 언급되고 있습니다. 그러나 메인 기사는 올바로 시스템을 통한 재생원료 인증의 구체적인 검증 과정에 초점을 맞추고 있는 반면, 비교 기사는 인증 시범사업이 국내 재활용사 및 배터리 제조사의 수출입 지원과 글로벌 무역규제 대응에 기여할 것이라는 점에 중점을 두고 있습니다."
+    },
+    {
+      "newsWithSimilarityDto": {
+        "similarity": 0.853,
+        "newsDto": {
+          "url": "https://www.mediapen.com/news/view/1026344",
+          "title": "“<b>폐배터리</b>에 미래 캔다&quot;…EV 배터리 재활용 시장 가파른 성장세",
+          "date": "2025-06-27",
+          "content": "사용 후 배터리의 처리와 재활용이 산업계와 정부 및 환경 분야에서 핵심 이슈로 부상하면서 순환경제 실현과 자원 확보, 글로벌 경쟁력 강화에 초점을 맞추며 관련 시장 육성에 박차를 가하고 있다.\n▲ SK에코플랜트 자회사 SK테스의 네덜란드 로테르담에 폐배터리 재활용 전처리 공장 모습./사진=SK에코플랜트 27일 시장조사기관 엑스퍼트마켓리서치에 따르면 국내 시장의 배터리 재활용 시장은 향후 10년 간 약 70%의 성장세를 보일 것으로 전망된다.\n글로벌 기준으로 전기차 배터리 재활용 시장은 올해 기준 약 49억 달러 규모로 평가되며 오는 2034년에는 428억 달러까지 성장세를 이어갈 것으로 예상된다.\n전문가들은 국내 사용후 배터리 재활용 시장이 2032년까지 3억4400만 달러(약 4655억 원) 규모로 성장할 것으로 보고 있다.\n연평균 성장률은 약 6%로 글로벌 시장의 고성장세(연평균 27.3%)에 비하면 다소 낮은 수치지만 국내 산업 구조와 정책 환경을 감안하면 의미 있는 성장세라는 평가다.\n전기차와 에너지저장장치(ESS) 등 친환경 모빌리티와 신재생에너지 산업이 확산되면서 수명이 다한 배터리의 재활용 수요도 함께 급증하고 있다.\n국내 배터리 재활용 시장의 성장 동력은 크게 △전기차 보급 확대 △환경 규제와 정책 강화 △자원 확보와 공급망 안정성 △기술 혁신과 인프라 확충 네 가지로 요약된다.\n전기차 보급 확대는 가장 큰 요인으로 평가된다.\n2024년 5월 기준 국내 전기차 누적 등록대수는 59만 대를 넘어섰다.\n앞서 환경부는 2030년 이후 연간 10만 개 이상의 사용후배터리가 배출될 것으로 내다봤다.\n이를 위해 환경부는 판매되는 전기차 외에도 이후 폐배터리 회수를 위해 배터리 환경성 계수를 적용하기도 했다.\n보조금에 차등 적용을 유인했으며 전기차 메이커들에게 핵심 광물 회수율이 높은 NCM(니켈, 코발트, 망간) 배터리 사용을 적극 권장했다.\n둘째는 환경 규제와 정책 강화다.\n정부는 2025년까지 재생원료 인증제를 도입하고 2027년부터 배터리 등급 분류를 위한 성능평가를 시행할 계획이다.\n해당 제도는 사용후배터리에서 추출한 리튬, 니켈 등 유가금속이 신품 배터리 제조에 얼마나 투입됐는지 확인하는 체계로 EU 등 주요국의 통상규제에 대응하고 국내 기업의 수출 경쟁력도 높일 것으로 기대된다.\n자원 확보와 공급망 안정성도 주효한 이유다.\n국내 배터리 산업은 리튬, 코발트, 니켈 등 핵심 광물의 대부분을 수입에 의존하고 있는 상황이다.\n배터리 재활용을 통해 금속들을 회수할 경우 원가 경쟁력과 공급망 안정성을 동시에 확보할 수 있다는 장점이 있다.\n환경부는 지난해 3월 국내 재활용 기업들과 ‘전기차 폐배터리 재생원료 인증 시범사업’ 협약을 체결했다.\n해당 사업은 폐배터리에서 추출한 재생원료의 품질과 이력을 인증하는 제도로 순환경제 실현과 희소 금속 공급망 강화에 기여하기 위한 취지다.\n▲ 성일하이텍 직원들이 폐 배터리 해체 작업을 하고 있다./사진=KBS1 시사멘터리 추적 '폐배터리가 몰려온다' 캡처 기술 혁신과 인프라 확충도 폐배터리 시장의 규모를 키우는 요인으로 꼽힌다.\n국내 주요 기업들은 습식제련, 건식제련, 직접 재활용 등 다양한 기술을 개발·적용하고 있다.\n성일하이텍, 에코프로, 포스코HY클린메탈, SK에코플랜트, 영풍, 고려아연 등은 배터리 전처리(분해·파쇄)부터 후처리(금속 추출·정제)까지 밸류체인을 내재화하며 경쟁력을 높이고 있다.\n특히 에코프로는 포항에 2조 원을 투입해 폐배터리 재활용 시스템을 구축했다.\n기업들의 노력을 장려하기 위해 정부의 지원도 확대되고 있다.\n환경부와 한국환경공단은 민간의 전기차 폐배터리 회수·관리체계 구축을 위해 분리, 보관, 화재방지, 성능평가, 방전 설비 구매비용의 50%를 지원하는 사업을 시행 중이다.\n사업자당 최대 1억 원의 지원금이 지급되며 이를 통해 민간 회수체계와 안전한 산업 환경 조성, 재사용·재활용 활성화가 기대된다.\n국내 배터리 재활용 산업은 글로벌 경쟁 심화, 규제 변화, 공급망 차질 등 과제도 안고 있다.\n그러나 정부와 업계의 협력, 기술 개발, 인프라 확충, 정책 지원이 맞물리며 시장은 지속적으로 성장할 것으로 전망된다.\n이를 위해 환경부는 생산자책임재활용제도(EPR) 대상을 기존 50종에서 전 품목으로 확대했다.\n전기·전자제품 내 배터리 회수율을 극대화하고 재활용이 까다로운 LFP(리튬, 인산, 철) 배터리 등도 EPR 대상에 포함해 공급망 차질에 적극 대응하고 있다.\n또한 전국 6개소에 있는 재활용 가능 자원 비축시설(약 1만5000톤 규모)을 원료 보관 장소로 기업에 임대해 원료 수급의 안정성을 높이고 폐기물 규제 완화로 순환자원 인정 범위도 확대하고 있다.\n환경부는 \"이 같은 대책을 계기로 국내 배터리 순환이용 업계가 향후 2~3년 내에 다가올 산업 성장기에도 시장 경제를 지속 유지할 수 있도록 적극 지원할 것\"이라고 말했다.",
+          "phrases": [
+            "전문가들은 국내 사용후 배터리 재활용 시장이 2032년까지 3억4400만 달러(약 4655억 원) 규모로 성장할 것으로 보고 있다.",
+            "정부는 2025년까지 재생원료 인증제를 도입하고 2027년부터 배터리 등급 분류를 위한 성능평가를 시행할 계획이다.",
+            "사용 후 배터리의 처리와 재활용이 산업계와 정부 및 환경 분야에서 핵심 이슈로 부상하면서 순환경제 실현과 자원 확보, 글로벌 경쟁력 강화에 초점을 맞추며 관련 시장 육성에 박차를 가하고 있다.",
+            "▲ SK에코플랜트 자회사 SK테스의 네덜란드 로테르담에 폐배터리 재활용 전처리 공장 모습./사진=SK에코플랜트 27일 시장조사기관 엑스퍼트마켓리서치에 따르면 국내 시장의 배터리 재활용 시장은 향후 10년 간 약 70%의 성장세를 보일 것으로 전망된다.",
+            "기업들의 노력을 장려하기 위해 정부의 지원도 확대되고 있다."
+          ]
+        }
+      },
+      "comparison": "메인 기사에서는 전기차 폐배터리 재생원료 인증 시범사업을 통해 인증제도 마련과 제도 개선을 추진하고 있으며, 이를 통해 순환경제 실현과 희소금속 공급망 안전성 강화를 목표로 하고 있습니다. 반면, 비교 기사에서는 전기차 보급 확대에 따른 배터리 재활용 시장의 급성장과 국내외 시장 전망에 초점을 맞추고 있습니다."
+    }
+  ]
+}
+                        }
                         }
                         }
                     }
@@ -145,11 +192,7 @@ async def analyze_news(dto: NewsDto):
     newsDtos: 비교 뉴스 리스트
     """
     try:
-        data = dto.dict()
-        if data.get("plan") == "PREMIUM":
-            return premium_analyze_and_summarize(data)
-        else:
-            return free_analyze_and_summarize(data)
+        return analyze_and_summarize(dto.dict())
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
